refactor(#14): interface specification according to its namespace and its builder specification
feature/specify-builder-methods-and-separate-interfaces-from-implementation
